Issue: WCSWREL-371 CLONE - Warnings in wifi_examples
[Issue]
Warnings in wifi_examples - unused variable and implicit conversion from float to double.

[Fix]
1. Variable is guarded inside macro to eliminate unused variable
warning.
2. Used float literal.

[Validation done]
Verified that warnings are not seen in compilation.

[Open Issues/Todo]
None

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sachin Patki

+++ wlcmgr/wlan_enhanced_tests.c
@@ -578,7 +578,7 @@ static void print_ds_rate(wlan_ds_rate ds_rate)
                     (void)PRINTF("    MCS:  Auto\r\n");
                 }
 
-                (void)PRINTF("    Rate: %f Mbps\r\n", (float)datarate->tx_data_rate / (float)2);
+                (void)PRINTF("    Rate: %f Mbps\r\n", (float)datarate->tx_data_rate / 2.0f);
             }
         }
 
@@ -621,7 +621,7 @@ static void print_ds_rate(wlan_ds_rate ds_rate)
                 {
                     (void)PRINTF("    MCS:  Auto\n");
                 }
-                (void)PRINTF("    Rate: %f Mbps\r\n", (float)datarate->rx_data_rate / (float)2);
+                (void)PRINTF("    Rate: %f Mbps\r\n", (float)datarate->rx_data_rate / 2.0f);
             }
         }
 #endif
